fixed hierarchy bugs

+++ frontend/src/features/projects/projectService.js
@@ -1,13 +1,15 @@
 import api from '../../utils/axiosInterceptor';
 
-const API_URL = 'http://localhost:3001/api/projects/'
-const getProject = async (projectId, token) => {
-    const response = await api.get(`${API_URL}/${projectId}`);
-    return response.data;
-}
+const API_URL = 'http://localhost:3001/api/projects'
+
 
 const getProjects = async (token) => {
     const response = await api.get(API_URL);
+    return response.data;
+}
+
+const getProject = async (projectId, token) => {
+    const response = await api.get(`${API_URL}/${projectId}`);
     return response.data;
 }
 
@@ -16,43 +18,81 @@ const createProject = async (projectData, token) => {
     return response.data;
 }
 
-const deleteProject = async (projectId, token) => {
-    const response = await api.delete(`${API_URL}/${projectId}`);
-    return response.data;
-}
-    
 const updateProject = async (projectId, projectData, token) => {
     const response = await api.put(`${API_URL}/${projectId}`, projectData);
     return response.data;
 }
 
+const deleteProject = async (projectId, token) => {
+    const response = await api.delete(`${API_URL}/${projectId}`);
+    return response.data;
+}
+
 const getHierarchy = async (projectId, token) => {
-    const response = await api.get(`${API_URL}${projectId}/hierarchy`)
-    return response.data
+    try {
+        const response = await api.get(`${API_URL}/${projectId}/hierarchy`)
+        return response.data
+    } catch (error) {
+        console.error('API call failed:', error);
+        throw error;
+    }
 }
 
-// Update hierarchy for a project
 const updateHierarchy = async (projectId, hierarchyData, token) => {
-    const response = await api.put(`${API_URL}${projectId}/hierarchy`, hierarchyData);
-    return response.data
+    const response = await api.put(`${API_URL}/${projectId}/hierarchy`, hierarchyData);
+    return response.data;
 }
 
-// Delete hierarchy for a project
 const deleteHierarchy = async (projectId, token) => {
-    const response = await api.delete(`${API_URL}${projectId}/hierarchy`);
-    return response.data
+    const response = await api.delete(`${API_URL}/${projectId}/hierarchy`);
+    return response.data;
 }
 
+const getHierarchyItemTypes = async (projectId, token) => {
+    const response = await api.get(`${API_URL}/${projectId}/hierarchy/item-types`);
+    return response.data;
+}
+
+const createHierarchyItemType = async (projectId, itemTypeData, token) => {
+    const response = await api.post(`${API_URL}/${projectId}/hierarchy/item-types`, itemTypeData);
+    return response.data;
+}
+
+const deleteHierarchyItemType = async (projectId, itemTypeId, token) => {
+    const response = await api.delete(`${API_URL}/${projectId}/hierarchy/item-types/${itemTypeId}`);
+    return response.data;
+}
+
+const getProjectUsers = async (projectId, token) => {
+    const response = await api.get(`${API_URL}/${projectId}/users`);
+    return response.data;
+}
+
+const addUserToProject = async (projectId, userData, token) => {
+    const response = await api.post(`${API_URL}/${projectId}/users`, userData);
+    return response.data;
+}
+
+const removeUserFromProject = async (projectId, userId, token) => {
+    const response = await api.delete(`${API_URL}/${projectId}/users/${userId}`);
+    return response.data;
+}
 
 const projectService = {
+    getProjects,
     getProject,
-    getProjects,        
     createProject,
-    deleteProject,  
     updateProject,
+    deleteProject,
     getHierarchy,
     updateHierarchy,
-    deleteHierarchy
-}   
+    deleteHierarchy,
+    getHierarchyItemTypes,
+    createHierarchyItemType,
+    deleteHierarchyItemType,
+    getProjectUsers,
+    addUserToProject,
+    removeUserFromProject
+}
 
-export default projectService
+export default projectService;
